// services/UI/projects/management-ui/src/app/api/model/menu.ts
/**
 * Mealist API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MenuPage } from './menuPage';


export interface Menu { 
    id?: string;
    name: string;
    status?: Menu.StatusEnum;
    dishesCount?: number;
    originalFileUrl?: string;
    stopColor?: string;
    stopStyle?: string;
    creationDate?: string;
    modificationDate?: string;
    pages?: Array<MenuPage>;
}
export namespace Menu {
    export type StatusEnum = 'NOT_PARSED' | 'PARSING_IN_PROGRESS' | 'PARSING_FAILED' | 'PARSING_COMPLETED' | 'OCR_IN_PROGRESS' | 'OCR_FAILED' | 'OCR_COMPLETED' | 'REVIEWED';
    export const StatusEnum = {
        NotParsed: 'NOT_PARSED' as StatusEnum,
        ParsingInProgress: 'PARSING_IN_PROGRESS' as StatusEnum,
        ParsingFailed: 'PARSING_FAILED' as StatusEnum,
        ParsingCompleted: 'PARSING_COMPLETED' as StatusEnum,
        OcrInProgress: 'OCR_IN_PROGRESS' as StatusEnum,
        OcrFailed: 'OCR_FAILED' as StatusEnum,
        OcrCompleted: 'OCR_COMPLETED' as StatusEnum,
        Reviewed: 'REVIEWED' as StatusEnum
    };
}