// services/UI/projects/management-ui/src/app/api/model/qrMenuItem.ts
/**
 * Mealist API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Menu } from './menu';


export interface QrMenuItem { 
    stopColor?: string;
    stopStyle?: string;
    title?: string;
    menu?: Menu;
}